pyproject.toml
[tool.poetry]
name = "monitapi"
version = "0.1.0"
description = "monitapi is something to monitor API(URL) by a serverless."
authors = ["fealone"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.61.1"
agraffe = "^0.2.0"
pydantic = "^1.6.1"
pyyaml = "^5.3.1"
aiohttp = {extras = ["speedups"], version = "^3.6.3"}


[tool.poetry.dev-dependencies]
uvicorn = {extras = ["watchdogreload"], version = "^0.12.1"}
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
